requirements.txt
flask
python-dotenv
embedchain